support saved games with large numbers of map indexes
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@4171 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Files/game_wad.cpp
@@ -2190,7 +2190,7 @@ static uint8 *tag_to_global_array_and_size(
 			count= dynamic_world->object_count;
 			break;
 		case MAP_INDEXES_TAG:
-			count= dynamic_world->map_index_count;
+			count= static_cast<unsigned short>(dynamic_world->map_index_count);
 			break;
 		case AUTOMAP_LINES:
 			count= (dynamic_world->line_count/8+((dynamic_world->line_count%8)?1:0));
